+++ src/modlunky2/sprites/base_classes/__init__.py
@@ -1,5 +1,11 @@
 from .base_sprite_loader import BaseSpriteLoader
 from .base_deco_sheet import AbstractDecoSheet
 from .base_biome import AbstractBiome
+from .base_floor_sheet import AbstractFloorSheet
 
-__all__ = ["BaseSpriteLoader", "AbstractDecoSheet", "AbstractBiome"]
+__all__ = [
+    "BaseSpriteLoader",
+    "AbstractDecoSheet",
+    "AbstractBiome",
+    "AbstractFloorSheet",
+]
